pallet_xp

Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

GenesisConfig

Genesis configuration for the XP pallet.

Instance1

Instance1 to be used for instantiable pallets defined with the #[pallet] macro. Instances 2-16 are also available but are hidden from docs.

Pallet

Primary Marker type for the XP pallet.

Enums

Call

XP Pallet Extrinsics includes major state mutation functions with origin authentication. Some read only functions are given for

Error

XP Pallet Errors

Event

XP Pallet Events (emitted via Pallet::deposit_event)

Traits

Config

Configuration trait of this pallet.

Type Aliases

InitXp

Stores the starting XP value of type Config::Xp for newly created XP keys.

LockedXpOf

Per-key locks (bounded by reason enum).

MinPulse

Minimum pulse required for XP heartbeat/reputation effects.

MinTimeStamp

Minimum timestamp (block number) for XP liveness.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

PulseFactor

Pulse factor parameters for XP activity reputation.

ReapedXp

Blacklist of finalized (reaped) XP keys.

ReservedXpOf

Per-key reserves.

XpOf

Stores XP state for key.

XpOwners

Owner-to-XP-key mapping.