pallet_xp::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

call

Executes a runtime call using an XP identity as the origin.

dispose

Dispose (Reap) an XP key.

force_genesis_config

Force-update a selected genesis configuration parameter.

force_handover

Force transfer/handover ownership of an XP key to another account.

handover

Transfer or handover ownership of an XP key to another account.

inspect_my_xp

Query the liquid XP balance for an owned XP key.

inspect_xp_keys_of

Emit a snapshot of all XpId’s currently owned by the specified account.