pallet_xp

Module types

Core types and aliases for the XP system.

This module defines the primary structures and type aliases used by pallet_xp. These types are publicly exposed and used across the pallet’s APIs for representing XP-related data.

Trait implementations provided by this crate’s Pallet can use these types via trait-bound equality constraints to ensure type alignment with this pallet’s concrete implementations if neccessary.

Example

mod pallet {
    use pallet_xp::types::Xp as XpData;

    pub trait Config<I: 'static>: frame_system::Config {
        type XpAdapter: XpSystem<Xp = XpData<Self, I>>;
    }
}

Structs

Accumulator

Internal accumulator structure for discrete XP pulse tracking.

GenesisAcc

Genesis configuration entry for an XP identity.

IdXp

A data structure that associates XP points with a specific reason identifier.

Stepper

Configuration structure for discrete accumulation operations.

Xp

The main XP data structure that is utilized on implementation of XpSystem::Xp.

XpProgress

Represents the progression mechanics behind XP scaling.

XpState

Snapshot of an identity’s XP-related state.

Enums

ForceGenesisConfig

Enumerates configurable XP parameters that may be forcibly overridden at runtime through privileged (root/governance) operations.

XpEligibility

Tracks an identity’s progress toward earning XP.

Type Aliases

LockReason

Reason identifier used when locking XP points.

ReserveReason

Reason identifier used when reserving XP points.

XpId

XP account identifier.

XpValue

Scalar XP value type representing the numerical XP amount.